Fix one small thing cuz mathias made getMaxTick return a long which is pretty wack cuz i didnt know you could split a tick in half wtf.
I did this.

+++ src/cs3500/easyanimator/controller/MVCController.java
@@ -19,7 +19,7 @@ public class MVCController implements IController {
   IVisualView view;
 
   private int tick;
-  private int tickRange;
+  private long tickRange;
   private Timer timer;
 
   private boolean shouldLoop;
